Filtering/pip-requirements.txt
cycler==0.10.0
decorator==4.4.2
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.1
matlabengineforpython===R2020b
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
progressbar==2.5
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tifffile==2021.4.8
